eyeguard/backend/requirements.txt
# Software-only simulation / demo — no real systems will be contacted or modified.
fastapi==0.110.0
uvicorn[standard]==0.27.1
httpx==0.27.0
redis==5.0.4
structlog==24.1.0
sqlalchemy==2.0.30
asyncpg==0.29.0
python-dotenv==1.0.1
reportlab==4.1.0
pandas==2.2.2
pytest==8.2.0
pytest-asyncio==0.23.5
pydantic-settings==2.11.0
email-validator==2.3.0
python-multipart==0.0.9
scapy==2.5.0
dnspython==2.4.2
